All the necessory changes are done
Constants, styled components, Jsx components are now separated
