+++ src/plays/hangman-game/styles.css
@@ -1,1 +1,2 @@
 /* enter stlyes here */
+
